+++ src/client/dialogs/CameraDialog.cpp
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-//    Scorched3D (c) 2000-2011, 2025
+//    Scorched3D (c) 2000-2011, 2025, 2026
 //
 //    This file is part of Scorched3D.
 //
@@ -124,11 +124,13 @@ void CameraDialog::draw()
 	
 	// Set the new viewport
 	targetCam_.getCamera().setWindowSize(
-		int(float(w_ - inset - inset) / GLViewPort::getWidthMult()), 
-		int(float(h_ - inset - inset) / GLViewPort::getHeightMult())); 
+		(int)( ( w_ - inset - inset ) / GLViewPort::getWidthMult() ),
+		(int)( ( h_ - inset - inset ) / GLViewPort::getHeightMult() )
+	);
 	targetCam_.getCamera().setWindowOffset(
-		int(float(x_ + inset) / GLViewPort::getWidthMult()), 
-		int(float(y_ + inset) / GLViewPort::getHeightMult()));
+		(int)( ( x_ + inset ) / GLViewPort::getWidthMult() ),
+		(int)( ( y_ + inset ) / GLViewPort::getHeightMult() )
+	);
 	targetCam_.draw();
 	
 	// Draw the stuff we want to see
