+++ src/client/graph/ShotCountDown.cpp
@@ -1,5 +1,5 @@
 ////////////////////////////////////////////////////////////////////////////////
-//    Scorched3D (c) 2000-2011, 2025
+//    Scorched3D (c) 2000-2011, 2025, 2026
 //
 //    This file is part of Scorched3D.
 //
@@ -140,8 +140,7 @@ void ShotCountDown::drawMove()
 		{
 			fontColor = &red;
 
-			str = S3D::formatStringBuffer("%.1f", 
-				move.timer_.asFloat());		
+			str = S3D::formatStringBuffer( "%.1f", move.timer_.asDouble() );
 		}
 		else
 		{
@@ -183,8 +182,7 @@ bool ShotCountDown::getRoundTime(std::string &str)
 	{
 		if (round.timer_ <= 5)
 		{
-			str = S3D::formatStringBuffer("%.1f", 
-				round.timer_.asFloat());		
+			str = S3D::formatStringBuffer( "%.1f", round.timer_.asDouble() );
 		}
 		else
 		{
